move code where it should live
