+++ src/main/kotlin/com/defenseunicorns/koci/Registry.kt
@@ -14,77 +14,7 @@ import io.ktor.http.*
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.channelFlow
 import kotlinx.coroutines.flow.flow
-import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
-import java.net.URI
-
-@Serializable
-data class CatalogResponse(val repositories: List<String>)
-
-@Serializable
-data class TagsResponse(val name: String, val tags: List<String>?)
-
-const val MANIFEST_MEDIA_TYPE = "application/vnd.oci.image.manifest.v1+json"
-const val MANIFEST_CONFIG_MEDIA_TYPE = "application/vnd.oci.image.config.v1+json"
-const val INDEX_MEDIA_TYPE = "application/vnd.oci.image.index.v1+json"
-
-private fun URLBuilder.paginate(n: Int, last: String? = null): URLBuilder = apply {
-    parameters.append("n", n.toString())
-    last?.let { parameters.append("last", it) }
-}
-
-class Router(registryURL: String) {
-    private val v2Prefix = "v2/"
-
-    private val base: URLBuilder = URLBuilder().takeFrom(registryURL).appendPathSegments(v2Prefix)
-
-    fun base(): Url {
-        return base.build()
-    }
-
-    fun catalog(): Url {
-        return base.clone().appendPathSegments("_catalog").build()
-    }
-
-    fun catalog(n: Int, lastRepo: String? = null): Url {
-        val catalogBase = URLBuilder().takeFrom(catalog())
-        return catalogBase.paginate(n, lastRepo).build()
-    }
-
-    fun tags(repository: String): Url {
-        return base.clone().appendPathSegments(repository, "tags", "list").build()
-    }
-
-    fun manifest(repository: String, ref: String): Url {
-        return base.clone().appendPathSegments(repository, "manifests", ref).build()
-    }
-
-    fun manifest(repository: String, descriptor: Descriptor): Url {
-        return manifest(repository, descriptor.digest.toString())
-    }
-
-    fun blob(repository: String, descriptor: Descriptor): Url {
-        return base.clone().appendPathSegments(repository, "blobs", descriptor.digest.toString())
-            .build()
-    }
-
-    fun uploads(repository: String): Url {
-        // the final "" allows for a trailing /
-        return base.clone().appendPathSegments(repository, "blobs", "uploads", "").build()
-    }
-
-    /**
-     * [RFC 7231](https://datatracker.ietf.org/doc/html/rfc7231#section-7.1.2)
-     */
-    fun parseUploadLocation(locationHeader: String): Url {
-        val uri = URI(locationHeader)
-        if (uri.isAbsolute) {
-            return URLBuilder().takeFrom(locationHeader).build()
-        }
-
-        return base.clone().appendPathSegments(locationHeader).build()
-    }
-}
 
 class Registry(
     registryURL: String,
@@ -204,11 +134,25 @@ class Registry(
     }
 }
 
+/**
+ * Connect to a repository.
+ */
 fun Registry.repo(name: String) = Repository(client, router, name)
+
+/**
+ * [GET /v2/<name>/tags/list](https://distribution.github.io/distribution/spec/api/#listing-image-tags)
+ */
 suspend fun Registry.tags(repository: String) = repo(repository).tags()
+
+/**
+ * [HEAD|GET /v2/<name>/manifests/<reference>](https://distribution.github.io/distribution/spec/api/#existing-manifests)
+ */
 suspend fun Registry.resolve(repository: String, tag: String, platformResolver: ((Platform) -> Boolean)? = null) =
     repo(repository).resolve(tag, platformResolver)
 
+/**
+ * Pull and tag.
+ */
 fun Registry.pull(
     repository: String,
     tag: String,
